Relaties lagen niet goed in gebied Product en ProductQuotations, deze bijgewerkt, testdata voor alle kolommen klaar, data adden OK, ready for UI

+++ KFSolutionsModel/NotMapped/OrderLine.cs
@@ -9,7 +9,7 @@ namespace KFSolutionsModel.NotMapped
     public class OrderLine : DBentity
     {
         public int NumberOfProducts { get; set; }
-        public int UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
 
 
     }
